+++ tmc-plugin-intellij/src/main/java/fi/helsinki/cs/tmc/intellij/actions/UploadExerciseAction.java
@@ -3,9 +3,9 @@ package fi.helsinki.cs.tmc.intellij.actions;
 import fi.helsinki.cs.tmc.intellij.holders.TmcCoreHolder;
 import fi.helsinki.cs.tmc.intellij.holders.TmcSettingsManager;
 import fi.helsinki.cs.tmc.intellij.io.CoreProgressObserver;
-import fi.helsinki.cs.tmc.intellij.services.CheckForExistingExercises;
-import fi.helsinki.cs.tmc.intellij.services.CourseAndExerciseManager;
-import fi.helsinki.cs.tmc.intellij.services.ExerciseUploadingService;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.CheckForExistingExercises;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.CourseAndExerciseManager;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.ExerciseUploadingService;
 import fi.helsinki.cs.tmc.intellij.services.ObjectFinder;
 import fi.helsinki.cs.tmc.intellij.services.ProgressWindowMaker;
 import fi.helsinki.cs.tmc.intellij.services.TestRunningService;
