+++ tmc-plugin-intellij/src/main/java/fi/helsinki/cs/tmc/intellij/holders/TmcSettingsManager.java
@@ -2,7 +2,7 @@ package fi.helsinki.cs.tmc.intellij.holders;
 
 
 import fi.helsinki.cs.tmc.intellij.io.SettingsTmc;
-import fi.helsinki.cs.tmc.intellij.services.PersistentTmcSettings;
+import fi.helsinki.cs.tmc.intellij.services.Persistence.PersistentTmcSettings;
 
 import com.intellij.openapi.components.ServiceManager;
