+++ tmc-plugin-intellij/src/main/java/fi/helsinki/cs/tmc/intellij/services/Errors/ErrorMessageService.java
@@ -1,4 +1,4 @@
-package fi.helsinki.cs.tmc.intellij.services;
+package fi.helsinki.cs.tmc.intellij.services.Errors;
 
 
 import fi.helsinki.cs.tmc.core.domain.Course;
@@ -17,6 +17,7 @@ import com.intellij.openapi.project.Project;
 
 import com.intellij.openapi.ui.Messages;
 
+import fi.helsinki.cs.tmc.intellij.services.ObjectFinder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
