+++ tmc-plugin-intellij/src/main/java/fi/helsinki/cs/tmc/intellij/services/Exercises/CourseAndExerciseManager.java
@@ -1,4 +1,4 @@
-package fi.helsinki.cs.tmc.intellij.services;
+package fi.helsinki.cs.tmc.intellij.services.Exercises;
 
 import fi.helsinki.cs.tmc.core.domain.Course;
 import fi.helsinki.cs.tmc.core.domain.Exercise;
@@ -9,16 +9,18 @@ import fi.helsinki.cs.tmc.intellij.holders.TmcCoreHolder;
 import fi.helsinki.cs.tmc.intellij.holders.TmcSettingsManager;
 import fi.helsinki.cs.tmc.intellij.io.SettingsTmc;
 
+import fi.helsinki.cs.tmc.intellij.services.Errors.ErrorMessageService;
+import fi.helsinki.cs.tmc.intellij.services.ObjectFinder;
+import fi.helsinki.cs.tmc.intellij.services.Persistence.ExerciseDatabase;
+import fi.helsinki.cs.tmc.intellij.services.Persistence.PersistentExerciseDatabase;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-// TODO Refactor CourseAndExerciseManager(), setup(), updateAll() -> initiateDatabase()!!!!
 /**
  * Holds a database of courses in memory, allowing quick fetching of course
  * when necessary without calling the TmcCore.
@@ -86,23 +88,7 @@ public class CourseAndExerciseManager {
             List<Course> courses = TmcCoreHolder.get()
                     .listCourses(ProgressObserver.NULL_OBSERVER).call();
 
-            for (Course course : courses) {
-                List<Exercise> exercises;
-                try {
-                    logger.info("Fetching {} from TmcCore. @CourseAndExerciseManager", course);
-                    course = TmcCoreHolder.get()
-                            .getCourseDetails(ProgressObserver.NULL_OBSERVER, course).call();
-                    exercises = new CheckForExistingExercises()
-                            .getListOfDownloadedExercises(course.getExercises(),
-                                    TmcSettingsManager.get());
-                    database.put(course.getName(), exercises);
-                } catch (Exception exception) {
-                    logger.warn("Failed to initiate database. @CourseAndExerciseManager",
-                            exception, exception.getStackTrace());
-                    new ErrorMessageService().showMessage(exception,
-                            "Failed to initiate database", true);
-                }
-            }
+            fetchCoursesFromTmcCore(database, courses);
 
             getDatabase().setCourses(database);
         } catch (TmcCoreException exception) {
@@ -117,6 +103,27 @@ public class CourseAndExerciseManager {
             ErrorMessageService error = new ErrorMessageService();
             error.showMessage(exception, "Failed to fetch courses from tmc core.", false);
             refreshCoursesOffline();
+        }
+    }
+
+    private void fetchCoursesFromTmcCore(Map<String, List<Exercise>> database, List<Course> courses) {
+        logger.info("Starting to fetch courses from TmcCore. @CourseAndExerciseManager");
+        for (Course course : courses) {
+            List<Exercise> exercises;
+            try {
+                logger.info("Fetching {} from TmcCore. @CourseAndExerciseManager", course);
+                course = TmcCoreHolder.get()
+                        .getCourseDetails(ProgressObserver.NULL_OBSERVER, course).call();
+                exercises = new CheckForExistingExercises()
+                        .getListOfDownloadedExercises(course.getExercises(),
+                                TmcSettingsManager.get());
+                database.put(course.getName(), exercises);
+            } catch (Exception exception) {
+                logger.warn("Failed to initiate database. @CourseAndExerciseManager",
+                        exception, exception.getStackTrace());
+                new ErrorMessageService().showMessage(exception,
+                        "Failed to initiate database", true);
+            }
         }
     }
