+++ tmc-plugin-intellij/src/main/java/fi/helsinki/cs/tmc/intellij/services/ObjectFinder.java
@@ -13,6 +13,7 @@ import com.intellij.openapi.actionSystem.DataKeys;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.project.ProjectManager;
 
+import fi.helsinki.cs.tmc.intellij.services.Errors.ErrorMessageService;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
