+++ tmc-plugin-intellij/src/main/java/fi/helsinki/cs/tmc/intellij/services/Persistence/ExerciseDatabase.java
@@ -1,4 +1,4 @@
-package fi.helsinki.cs.tmc.intellij.services;
+package fi.helsinki.cs.tmc.intellij.services.Persistence;
 
 import fi.helsinki.cs.tmc.core.domain.Exercise;
