+++ tmc-plugin-intellij/src/main/java/fi/helsinki/cs/tmc/intellij/spyware/ButtonInputListener.java
@@ -2,7 +2,7 @@ package fi.helsinki.cs.tmc.intellij.spyware;
 
 
 import fi.helsinki.cs.tmc.core.domain.Exercise;
-import fi.helsinki.cs.tmc.intellij.services.CourseAndExerciseManager;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.CourseAndExerciseManager;
 import fi.helsinki.cs.tmc.intellij.services.ObjectFinder;
 import fi.helsinki.cs.tmc.intellij.services.PathResolver;
 import fi.helsinki.cs.tmc.spyware.LoggableEvent;
