+++ tmc-plugin-intellij/src/main/java/fi/helsinki/cs/tmc/intellij/spyware/SpywareFileListener.java
@@ -3,7 +3,7 @@ package fi.helsinki.cs.tmc.intellij.spyware;
 import fi.helsinki.cs.tmc.core.domain.Exercise;
 import fi.helsinki.cs.tmc.core.utilities.JsonMaker;
 import fi.helsinki.cs.tmc.intellij.holders.TmcSettingsManager;
-import fi.helsinki.cs.tmc.intellij.services.CourseAndExerciseManager;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.CourseAndExerciseManager;
 import fi.helsinki.cs.tmc.intellij.services.PathResolver;
 import fi.helsinki.cs.tmc.intellij.spyware.spywareutils.ActiveThreadSet;
 import fi.helsinki.cs.tmc.intellij.spyware.spywareutils.RecursiveZipper;
