+++ tmc-plugin-intellij/src/main/java/fi/helsinki/cs/tmc/intellij/ui/pastebin/SubmitPanel.java
@@ -1,7 +1,7 @@
 package fi.helsinki.cs.tmc.intellij.ui.pastebin;
 
 import fi.helsinki.cs.tmc.intellij.holders.ProjectListManagerHolder;
-import fi.helsinki.cs.tmc.intellij.services.CourseAndExerciseManager;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.CourseAndExerciseManager;
 import fi.helsinki.cs.tmc.intellij.services.PasteService;
 
 import com.intellij.uiDesigner.core.GridConstraints;
