+++ tmc-plugin-intellij/src/main/java/fi/helsinki/cs/tmc/intellij/ui/projectlist/ProjectListManager.java
@@ -1,7 +1,7 @@
 package fi.helsinki.cs.tmc.intellij.ui.projectlist;
 
 import fi.helsinki.cs.tmc.core.domain.Exercise;
-import fi.helsinki.cs.tmc.intellij.services.CourseAndExerciseManager;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.CourseAndExerciseManager;
 import fi.helsinki.cs.tmc.intellij.services.ObjectFinder;
 
 import com.intellij.openapi.application.ApplicationManager;
