+++ tmc-plugin-intellij/src/test/java/fi/helsinki/cs/tmc/intellij/services/ExerciseDownloadingServiceTest.java
@@ -5,12 +5,13 @@ import com.intellij.openapi.project.Project;
 import fi.helsinki.cs.tmc.core.TmcCore;
 import fi.helsinki.cs.tmc.intellij.io.CoreProgressObserver;
 import fi.helsinki.cs.tmc.intellij.io.SettingsTmc;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.CheckForExistingExercises;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.ExerciseDownloadingService;
 import org.junit.Test;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.when;
 
 public class ExerciseDownloadingServiceTest {
