+++ tmc-plugin-intellij/src/test/java/fi/helsinki/cs/tmc/intellij/services/PasteServiceTest.java
@@ -4,10 +4,10 @@ package fi.helsinki.cs.tmc.intellij.services;
 import fi.helsinki.cs.tmc.core.TmcCore;
 import fi.helsinki.cs.tmc.core.domain.Exercise;
 import fi.helsinki.cs.tmc.core.domain.ProgressObserver;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.CourseAndExerciseManager;
 import fi.helsinki.cs.tmc.intellij.ui.pastebin.PasteWindow;
 import fi.helsinki.cs.tmc.intellij.ui.projectlist.ProjectListManager;
 import org.junit.Test;
-import org.mockito.Mockito;
 
 import java.net.URI;
 import java.util.concurrent.Callable;
