+++ tmc-plugin-intellij/src/test/java/fi/helsinki/cs/tmc/intellij/services/PersistentTmcSettingsTest.java
@@ -5,6 +5,7 @@ import static org.junit.Assert.assertNotEquals;
 
 import fi.helsinki.cs.tmc.intellij.io.SettingsTmc;
 
+import fi.helsinki.cs.tmc.intellij.services.Persistence.PersistentTmcSettings;
 import org.junit.Before;
 import org.junit.Test;
