+++ tmc-plugin-intellij/src/test/java/fi/helsinki/cs/tmc/intellij/services/UploadExerciseServiceTest.java
@@ -8,6 +8,9 @@ import fi.helsinki.cs.tmc.core.TmcCore;
 import fi.helsinki.cs.tmc.core.domain.Course;
 import fi.helsinki.cs.tmc.intellij.io.CoreProgressObserver;
 import fi.helsinki.cs.tmc.intellij.io.SettingsTmc;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.CheckForExistingExercises;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.CourseAndExerciseManager;
+import fi.helsinki.cs.tmc.intellij.services.Exercises.ExerciseUploadingService;
 import fi.helsinki.cs.tmc.intellij.ui.submissionresult.SubmissionResultHandler;
 import org.junit.Rule;
 import org.junit.Test;
